sql: fix seahash catalog return type (uint4 -> uint8) (#36360)
The seahash functions return u64 but were registered with UInt32 in
func.rs, causing mz_functions.return_type_id to disagree with the actual
runtime type. Introduced in PR #29099.

+++ src/sql/src/func.rs
@@ -4759,9 +4759,9 @@ pub static MZ_CATALOG_BUILTINS: LazyLock<BTreeMap<&'static str, Func>> = LazyLoc
         },
         "seahash" => Scalar {
             params!(String) => UnaryFunc::SeahashString(func::SeahashString)
-                => UInt32, oid::FUNC_SEAHASH_STRING_OID;
+                => UInt64, oid::FUNC_SEAHASH_STRING_OID;
             params!(Bytes) => UnaryFunc::SeahashBytes(func::SeahashBytes)
-                => UInt32, oid::FUNC_SEAHASH_BYTES_OID;
+                => UInt64, oid::FUNC_SEAHASH_BYTES_OID;
         },
         "starts_with" => Scalar {
             params!(String, String) => BinaryFunc::from(func::StartsWith) => Bool, 3696;
